neo4j_data_ingestion/requirements.txt
APScheduler==3.5.1
certifi==2017.11.5
chardet==3.0.4
funcsigs==1.0.2
futures==3.2.0
idna==2.6
MySQL-python==1.2.5
neo4j-driver==1.5.3
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
